Move Exercise.java and ExerciseList.java to data.exercises package
